+++ src/app/mork/celery/tasks/emailing.py
@@ -19,19 +19,23 @@ logger = getLogger(__name__)
 
 
 @app.task
-def warn_inactive_users(dry_run: bool = True):
+def warn_inactive_users(limit: int = 0, dry_run: bool = True):
     """Celery task to warn inactive users by email."""
     db = OpenEdxMySQLDB()
 
     threshold_date = datetime.now() - settings.WARNING_PERIOD
-
     total = crud.get_inactive_users_count(db.session, threshold_date)
+
+    if limit:
+        total = min(total, limit)
+
     for batch_offset in range(0, total, settings.EDX_MYSQL_QUERY_BATCH_SIZE):
+        batch_limit = min(settings.EDX_MYSQL_QUERY_BATCH_SIZE, total - batch_offset)
         inactive_users = crud.get_inactive_users(
             db.session,
             threshold_date,
             offset=batch_offset,
-            limit=settings.EDX_MYSQL_QUERY_BATCH_SIZE,
+            limit=batch_limit,
         )
         send_email_group = group(
             [